Some fixes

+++ app/src/main/java/com/brewthings/app/ui/components/graph/GraphState.kt
@@ -9,6 +9,6 @@ data class GraphSeries (
 )
 
 data class GraphDataPoint (
-    val x: Long,
+    val x: Float,
     val y: Float
 )
